Move normalized cache classes to cache/normalized
