+++ apollo-integration/src/test/java/com/apollographql/android/impl/ResponseNormalizationTest.java
@@ -6,10 +6,10 @@ import com.google.common.io.Files;
 import com.apollographql.android.ApolloCall;
 import com.apollographql.android.CustomTypeAdapter;
 import com.apollographql.android.api.graphql.Response;
-import com.apollographql.android.cache.Record;
-import com.apollographql.android.cache.CacheKeyResolver;
-import com.apollographql.android.cache.CacheReference;
-import com.apollographql.android.cache.InMemoryCacheStore;
+import com.apollographql.android.cache.normalized.Record;
+import com.apollographql.android.cache.normalized.CacheKeyResolver;
+import com.apollographql.android.cache.normalized.CacheReference;
+import com.apollographql.android.cache.normalized.InMemoryCacheStore;
 import com.apollographql.android.impl.normalizer.EpisodeHeroName;
 import com.apollographql.android.impl.normalizer.HeroAndFriendsNames;
 import com.apollographql.android.impl.normalizer.HeroAndFriendsNamesWithIDForParentOnly;
@@ -18,7 +18,6 @@ import com.apollographql.android.impl.normalizer.HeroAppearsIn;
 import com.apollographql.android.impl.normalizer.HeroName;
 import com.apollographql.android.impl.normalizer.HeroTypeDependentAliasedField;
 import com.apollographql.android.impl.normalizer.SameHeroTwice;
-import com.apollographql.android.impl.normalizer.type.Episode;
 
 import org.junit.Before;
 import org.junit.Test;
