+++ apollo-runtime/src/main/java/com/apollographql/android/cache/normalized/Cache.java
@@ -1,4 +1,4 @@
-package com.apollographql.android.cache;
+package com.apollographql.android.cache.normalized;
 
 import com.apollographql.android.impl.ResponseNormalizer;
