+++ apollo-runtime/src/main/java/com/apollographql/android/cache/normalized/CacheKeyResolver.java
@@ -1,4 +1,4 @@
-package com.apollographql.android.cache;
+package com.apollographql.android.cache.normalized;
 
 import java.util.Map;
