+++ apollo-runtime/src/main/java/com/apollographql/android/impl/ApolloClient.java
@@ -8,9 +8,9 @@ import com.apollographql.android.api.graphql.ScalarType;
 import com.apollographql.android.cache.http.EvictionStrategy;
 import com.apollographql.android.cache.http.HttpCache;
 import com.apollographql.android.cache.http.ResponseCacheStore;
-import com.apollographql.android.cache.Cache;
-import com.apollographql.android.cache.CacheKeyResolver;
-import com.apollographql.android.cache.CacheStore;
+import com.apollographql.android.cache.normalized.Cache;
+import com.apollographql.android.cache.normalized.CacheKeyResolver;
+import com.apollographql.android.cache.normalized.CacheStore;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonWriter;
 import com.squareup.moshi.Moshi;
